pyproject.toml
[project]
name = "ktds-msai-7th"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-search-documents>=11.6.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
